+++ src/main/java/com/kuit/moamoa/dto/request/chat/ChatMessageRequest.java
@@ -1,6 +1,5 @@
-package com.kuit.moamoa.dto.request;
+package com.kuit.moamoa.dto.request.chat;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,7 +9,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UpdateChatRequest {
-    @NotBlank(message = "Content cannot be blank")
+public class ChatMessageRequest {
+    private Long userGroupId;
+    private Long userId;
     private String content;
 }
