+++ src/main/java/com/kuit/moamoa/dto/request/chat/UpdateChatRequest.java
@@ -1,5 +1,6 @@
-package com.kuit.moamoa.dto.request;
+package com.kuit.moamoa.dto.request.chat;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChatMessageRequest {
-    private Long userGroupId;
-    private Long userId;
+public class UpdateChatRequest {
+    @NotBlank(message = "Content cannot be blank")
     private String content;
 }
