+++ src/main/java/com/kuit/moamoa/service/ChatService.java
@@ -4,9 +4,9 @@ import com.kuit.moamoa.domain.Chat;
 import com.kuit.moamoa.domain.Status;
 import com.kuit.moamoa.domain.User;
 import com.kuit.moamoa.domain.UserGroup;
-import com.kuit.moamoa.dto.request.ChatDeleteEvent;
-import com.kuit.moamoa.dto.request.ChatMessageRequest;
-import com.kuit.moamoa.dto.response.ChatMessageResponse;
+import com.kuit.moamoa.dto.request.chat.ChatDeleteEvent;
+import com.kuit.moamoa.dto.request.chat.ChatMessageRequest;
+import com.kuit.moamoa.dto.response.chat.ChatMessageResponse;
 import com.kuit.moamoa.global.exception.ChatException;
 import com.kuit.moamoa.global.exception.ErrorCode;
 import com.kuit.moamoa.global.response.ApiResponse;
